Removed int enum casting in unit test

+++ OOP Assignment 2/Group40/UnitTest1/Assignment_UnitTests.cpp
@@ -221,7 +221,7 @@ namespace Conflict_UnitTests
 
 			//Act
 			brawler.Defend(1); //invalid index! no armour in vector
-			actualState = static_cast<int>(brawler.GetState());
+			actualState = brawler.GetState();
 
 			//Assert
 			Assert::AreEqual(expectedState, actualState);
